# models.py
# generated by fastapi-codegen:
#   filename:  openapi.yaml
#   timestamp: 2024-06-04T22:41:01+00:00

from __future__ import annotations

from datetime import datetime
from typing import List, Optional

from sqlmodel import SQLModel, Field, Relationship

class Error(SQLModel):
    message: str
    timestamp: Optional[datetime] = None


class Healthcheck(SQLModel):
    message: str
    database_status: str
    timestamp: datetime


class Cart(SQLModel, table=True):
    id: Optional[int] = Field(default=None, primary_key=True)
    name: str
    items: list["ItemsInCart"] = Relationship(back_populates="cart")


class Item(SQLModel, table=True):
    id: Optional[int] = Field(default=None, primary_key=True)
    name: str

    carts: list["ItemsInCart"] = Relationship(back_populates="item")

class ItemsInCart(SQLModel, table=True):
    id: Optional[int] = Field(default=None, primary_key=True)
    cart_id: int = Field(foreign_key="cart.id")
    cart: Cart = Relationship(back_populates="items")
    item_id: int = Field(foreign_key="item.id")
    item: Item = Relationship(back_populates="carts")

class ItemQty(SQLModel):
    item: Item
    quantity: int


class CartSummary(SQLModel):
    cart: Cart
    items: List[ItemQty]